feat(orga): moving campaign participation result repository
