+++ api/src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-result-repository.js
@@ -1,10 +1,10 @@
-import { CampaignParticipationResult } from '../../../src/shared/domain/models/CampaignParticipationResult.js';
-import * as areaRepository from '../../../src/shared/infrastructure/repositories/area-repository.js';
-import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
-import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as campaignParticipationRepository from './../../../src/prescription/campaign-participation/infrastructure/repositories/campaign-participation-repository.js';
-import * as campaignRepository from './campaign-repository.js';
-import * as knowledgeElementRepository from './knowledge-element-repository.js';
+import * as campaignRepository from '../../../../../lib/infrastructure/repositories/campaign-repository.js';
+import * as knowledgeElementRepository from '../../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
+import { CampaignParticipationResult } from '../../../../shared/domain/models/CampaignParticipationResult.js';
+import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
+import * as assessmentRepository from '../../../../shared/infrastructure/repositories/assessment-repository.js';
+import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
+import * as campaignParticipationRepository from './campaign-participation-repository.js';
 
 const campaignParticipationResultRepository = {
   async getByParticipationId(campaignParticipationId) {
